backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.40.0
python-dotenv==1.0.1
